pyproject.toml
[project]
name = "goodreads-bot-reloaded"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "gql[all]>=3.5.0",
    "praw>=7.8.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
